Fixed: container wait() timeout not working
The problem is that the context where the timeout applies was only the
connection. Therefore the response.read() call hangs until it reaches
the ClientSession's timeout.

I believe the timeout worked before depending on the conditions. If the
timeout value is pretty low but the server is pretty fast, the timeout
may raises when the connection initializes but before the read()
actually happen.