Implement the ability to use docker exec.
This builds on the great work in
https://github.com/hirokiky/aiodocker/commits/feature-exec-run. It
improves how stdout and stderr are handled so that, when a Tty is not
used, we can handle the multiplexing that is done by the docker daemon
when writing from the websocket. It includes information on whether the
data is from stdout or stderr. We can include that information in the
messages we create.
If a Tty is used, it treats all data as coming from stdout because
docker can not make any distinction between streams in that case.

We can also get rid of the `stream` argument and just use `Detach`. If
`Detach` is True, we just return with any output of doing that, which is
probably (maybe always) going to be empty.
If `Detach` is False, then we'll create the ClientWebSocketResponse
object that can be used to read from/write to the process.